+++ changelog.md
@@ -22,6 +22,7 @@ to pATLAS version 1.5.2.
 
 - Added check for `params.accessions` that enables to report a proper
 error when it is set to `null`.
+- Added `build` option to export component parameters information in JSON format. 
 
 ### Bug fixes